Compute branch displacement narrowing in 64 bits
check_oversized_branch added size_delta to the rel32 displacement in
int32_t, which overflows -- undefined behavior -- for displacements
within 3-4 bytes of INT32_MAX, i.e. a branch spanning a ~2 GB section.
Under the usual wrapping semantics the result happened to come out
right (the wrapped sum lands far outside rel8 range, so no finding),
but a UBSan build traps on it: jmp +0x7ffffffd reproduced.

Do the sum in int64_t. The verdict at the boundary is unchanged and now
defined; tests pin both overflowing encodings (JMP +0x7ffffffd and
Jcc +0x7ffffffc).

+++ x86lint.c
@@ -1229,8 +1229,10 @@ bool check_oversized_branch(const xed_decoded_inst_t *xedd)
         return true;
     }
 
+    // Computed in 64 bits: a rel32 displacement near INT32_MAX (a branch
+    // spanning a ~2 GB section) would overflow the int32 sum.
     int32_t disp = (int32_t) xed_decoded_inst_get_branch_displacement(xedd);
-    int32_t new_disp = disp + size_delta;
+    int64_t new_disp = (int64_t) disp + size_delta;
     return new_disp < INT8_MIN || new_disp > INT8_MAX;
 }
